// src/infrastructure/third-party/notebooklm/rpc/types.ts
export const QUERY_URL =
  "https://notebooklm.google.com/_/LabsTailwindUi/data/google.internal.labs.tailwind.orchestration.v1.LabsTailwindOrchestrationService/GenerateFreeFormStreamed";

export const UPLOAD_URL = "https://notebooklm.google.com/upload/_/";

export enum RPCMethod {
  LIST_NOTEBOOKS = "wXbhsf",
  CREATE_NOTEBOOK = "CCqFvf",
  GET_NOTEBOOK = "rLM1Ne",
  RENAME_NOTEBOOK = "s0tc2d",
  DELETE_NOTEBOOK = "WWINqb",

  ADD_SOURCE = "izAoDd",
  ADD_SOURCE_FILE = "o4cbdc",
  DELETE_SOURCE = "tGMBJ",
  GET_SOURCE = "hizoJc",
  REFRESH_SOURCE = "FLmJqe",
  CHECK_SOURCE_FRESHNESS = "yR9Yof",
  UPDATE_SOURCE = "b7Wfje",
  DISCOVER_SOURCES = "qXyaNe",

  SUMMARIZE = "VfAZjd",
  GET_SOURCE_GUIDE = "tr032e",
  GET_SUGGESTED_REPORTS = "ciyUvf",

  CREATE_ARTIFACT = "R7cb6c",
  LIST_ARTIFACTS = "gArtLc",
  DELETE_ARTIFACT = "V5N4be",
  RENAME_ARTIFACT = "rc3d8d",
  EXPORT_ARTIFACT = "Krh3pd",
  SHARE_ARTIFACT = "RGP97b",
  GET_INTERACTIVE_HTML = "v9rmvd",
  REVISE_SLIDE = "KmcKPe",

  START_FAST_RESEARCH = "Ljjv0c",
  START_DEEP_RESEARCH = "QA9ei",
  POLL_RESEARCH = "e3bVqc",
  IMPORT_RESEARCH = "LBwxtb",

  GENERATE_MIND_MAP = "yyryJe",
  CREATE_NOTE = "CYK0Xb",
  GET_NOTES_AND_MIND_MAPS = "cFji9",
  UPDATE_NOTE = "cYAfTb",
  DELETE_NOTE = "AH0mwd",

  GET_LAST_CONVERSATION_ID = "hPTbtc",
  GET_CONVERSATION_TURNS = "khqZz",

  SHARE_NOTEBOOK = "QDyure",
  GET_SHARE_STATUS = "JFMDGd",

  REMOVE_RECENTLY_VIEWED = "fejl7e",

  GET_USER_SETTINGS = "ZwVcOc",
  SET_USER_SETTINGS = "hT54vc",
}
// ...
export enum ChatGoal {
  DEFAULT = 1,
  CUSTOM = 2,
  LEARNING_GUIDE = 3,
}

export enum ChatResponseLength {
  DEFAULT = 1,
  LONGER = 4,
  SHORTER = 5,
}